Intelligently handle document-local links
Rewrite href attributes of "a" elements if their targets is within
the captured fragment but they are referenced by full path
or even by full URI.

Make local links work in the captured document even if hash
sign is in the middle of the href attribute.

In the case of page fragment capture, distinguish link targets
within the selection and outside of it.

Caveats:
* In the case of partial capture deep in the nested
page structure, the links to ancestors are considered local
however ancestors content is not included.
* Obsolete <A name="bla-bla-bla"> anchors are not handled
* In the case of page fragment selection and a really old firefox
  version (Selector API is not supported) all links are rewritten
  to the original URI.