capture: fix error for partial capture with self link
If a partial contains a link to self not contained in the captured region, a full copy of index.html will be saved (probably default.html) and will be written to gURL2Name causing index not in gURL2Name. This causes the index be not in sbCrossLinker.nameList and subsequent script error. We prepend "index" to sbCrossLinker.nameList if it doesn't exist to fix this issue.