capture: fix over-encoding for xhtml
"!", "'", "(", ")", and "~" are reserved URL chars which should not be encoded in an URL. However, when setting the href attribute of an anchor to a URL containing these chars and save to a file, they are saved encoded as "%21", "%27", "%28", "%29", and "%7E".

We force gURL2Name save and compare using a normalized form to correct this issue.