capture: fix url with css escape sequence
The .cssText property value Firefox returns usually has escaped sequences, such as "\0000af " and "\#", unescaped. However, in certain cases the returned value does contain the escape sequence. For example, "url("A's.png")" is returned as "url("A\'s.png")" on Firefox 50.

We add an unescapeCss() when we are going to retrieve the text value from CSS to solve the issue.