capture: convert xhtml code to html if fileAsHtml
If we directly save the XHTML output into an .html file, something may not work right when it's opened.

For example, <style> and <script> tags in XHTML are strict XML tags and "&" must be written as "&amp;", while in HTML they are text and "&" should be written as-is. If we save XHTML output to HTML, an "&amp;" will be regarded as-is and corrupts the related URL.