save: remove hash in gURL2Name
URLs that are different only in hash are now regarded as identical in gURL2Name.

This fixes an error, for example, save a page with URL "index.html#123" that links to another page "linked.html" which has a link to "index.html", and the link in "linked.html" was not rewritten.